Started adding closures
