+++ src/closures.c
@@ -1,0 +1,12 @@
+/*
+ * File: closures.c
+ * ----------------
+ * Presents the implementation of closure related functionality
+ */
+
+#include "closures.h"
+
+obj* make_closure(const obj* lambda, obj** envp) {
+  // todo
+  return NULL;
+}
